feat: Add Candidates List Component with Filtering, Sorting, and Pagination
- Implemented CandidatesList component to display candidates with details.
- Added loading and error handling states.
- Integrated Redux for managing candidates' state, including filters and pagination.
- Created CandidateCard component for individual candidate display with status update functionality.
- Introduced FiltersModal for applying filters on candidates.
- Enhanced pagination functionality to navigate through candidates.
- Updated candidatesSlice to include new fields and handle fetching candidates with additional data.
- Modified supabase types to accommodate new candidate fields.
