+++ src/app/jobs/page.tsx
@@ -1,3 +1,4 @@
+"use server";
 import { createClient } from "@/utils/supabase/server"; // Use server client
 import JobsClientComponent from "./JobsClientComponent";
 
@@ -21,7 +22,7 @@ import JobsClientComponent from "./JobsClientComponent";
 //   }
 // }
 
-export const revalidate = 1800; // 30 minutes
+// export const revalidate = 1800; // 30 minutes
 
 export default async function JobsPage() {
   // Fetch jobs on the server side
